# docs/technical/phase8/OpenAPI_Specification.yaml
openapi: 3.1.0
info:
  title: Majhi Dairy API
  version: 1.0.0
  description: API contract for Majhi Dairy frontend, Supabase backend, AI, OCR and admin integrations.
servers:
  - url: https://api.majhidairy.app
    description: Production API base
  - url: http://localhost:3000
    description: Local Next.js API base
security:
  - bearerAuth: []
...
paths:
  "/v1/auth/signup":
    post:
      tags:
        - Authentication
      operationId: AUTH_API_001
      summary: "Create user, profile, first farm and language onboarding state."
      description: "name plus email or phone required; password policy; language mr/en"
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericRequest'
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
        "422":
          $ref: '#/components/responses/ValidationFailed'
        "429":
          $ref: '#/components/responses/RateLimited'
        "500":
          $ref: '#/components/responses/InternalError'
  "/v1/auth/login":
    post:
      tags:
        - Authentication
      operationId: AUTH_API_002
      summary: "Authenticate by email/phone and password."
      description: "identifier and password required"
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericRequest'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
        "422":
          $ref: '#/components/responses/ValidationFailed'
        "429":
          $ref: '#/components/responses/RateLimited'
        "500":
          $ref: '#/components/responses/InternalError'
...
  "/v1/support/faq":
    get:
      tags:
        - Support
      operationId: SUP_API_009
      summary: "Search FAQ and help articles."
      description: "language mr/en"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
        "422":
          $ref: '#/components/responses/ValidationFailed'
        "429":
          $ref: '#/components/responses/RateLimited'
        "500":
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StandardResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          additionalProperties: true
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        message:
          type: string
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            localizedMessage:
              type: object
              properties:
                mr:
                  type: string
                en:
                  type: string
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
            details:
              type: object
              additionalProperties: true
        requestId:
          type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        hasNext:
          type: boolean
        requestId:
          type: string
    GenericRequest:
      type: object
      additionalProperties: true
...
  responses:
    BadRequest:
      description: Bad request
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Unauthorized:
      description: Authentication required
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Forbidden:
      description: Access denied
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Conflict:
      description: Conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    ValidationFailed:
      description: Validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    RateLimited:
      description: Rate limited
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    InternalError:
      description: Internal server error
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }